read

+++ 0x15-file_io/main.h
@@ -1,5 +1,5 @@
 #ifndef MAIN_H
 #define MAIN_H
 int _putchar(char c);
-int read_textfile(char *filename, size_t letters);
+int read_textfile(char *filename, int letters);
 #endif
